fix bug related to type of bert_warmup option

+++ onmt/opts.py
@@ -466,7 +466,7 @@ def train_opts(parser):
     group.add('--warmup_steps', '-warmup_steps', type=int, default=4000,
               help="""Number of warmup steps for custom decay.""")
 
-    group.add('--bert_warmup', '-bert_warmup', type=int, default=0.1,
+    group.add('--bert_warmup', '-bert_warmup', type=float, default=0.1,
               help="""portion of total steps for the warmup.""")
 
     group = parser.add_argument_group('Logging')
